+++ Endpoints/Admin/AdminEndpoint.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Newtonsoft.Json.Linq;
+using SwarmUI.ApiClient.Contracts.Responses;
 using SwarmUI.ApiClient.Http;
-using SwarmUI.ApiClient.Models.Responses;
 using SwarmUI.ApiClient.Sessions;
 
 namespace SwarmUI.ApiClient.Endpoints.Admin;