+++ Endpoints/User/IUserEndpoint.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using SwarmUI.ApiClient.Models.Responses;
+using SwarmUI.ApiClient.Contracts.Responses;
 
 namespace SwarmUI.ApiClient.Endpoints.User;