+++ Tests/IntegrationTests/BackendsIntegrationTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using SwarmUI.ApiClient.Models.Responses;
+using SwarmUI.ApiClient.Contracts.Responses;
 using Xunit;
 
 namespace SwarmUI.ApiClient.Tests.IntegrationTests