+++ Tests/IntegrationTests/GenerationIntegrationTests.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using SwarmUI.ApiClient.Models.Common;
-using SwarmUI.ApiClient.Models.Requests;
-using SwarmUI.ApiClient.Models.Responses;
+using SwarmUI.ApiClient.Contracts.Common;
+using SwarmUI.ApiClient.Contracts.Requests;
+using SwarmUI.ApiClient.Contracts.Responses;
 using Xunit;
 
 namespace SwarmUI.ApiClient.Tests.IntegrationTests